+++ jai/__init__.py
@@ -5,10 +5,10 @@
 __version__ = "v0.19.1"
 
 from jai.core.jai import Jai
-from jai.utilities._image import read_image
-from jai.utilities._processing import (filter_resolution, filter_similar,
+from jai.utilities.image import read_image
+from jai.utilities.processing import (filter_resolution, filter_similar,
                                        find_threshold, predict2df, treat_unix)
-from jai.utilities._splits import split, split_recommendation
+from jai.utilities.splits import split, split_recommendation
 
 __all__ = [
     "__version__", "Jai", "read_image", "split", "split_recommendation",
