+++ jai/tests/test_splits.py
@@ -2,7 +2,7 @@ import pandas as pd
 import pytest
 from pandas.testing import assert_frame_equal
 
-from jai.utilities._splits import split
+from jai.utilities.splits import split
 
 
 # =============================================================================
