+++ jai/utilities/__init__.py
@@ -1,7 +1,7 @@
-from ._image import read_image
-from ._processing import (filter_resolution, filter_similar, find_threshold,
-                          predict2df, treat_unix)
-from ._splits import split, split_recommendation
+from .image import read_image
+from .processing import (filter_resolution, filter_similar, find_threshold,
+                         predict2df, treat_unix)
+from .splits import split, split_recommendation
 
 __all__ = [
     "read_image", "split", "split_recommendation", "find_threshold",
