+++ jai/utilities/image.py
